requirements.txt
pdfplumber
python-dotenv
openai
tqdm
